refactor(textArea): Refactor TextAreaModule component's function, ref, state, side effect for better division (#14)
