+++ src/components/Input/TextArea/TextAreaModule.jsx
@@ -1,4 +1,6 @@
 /* eslint-disable react/prop-types */
+import { useEffect, useRef, useState } from "react";
+
 import {
   TextArea,
   TextAreaCharacterCount,
@@ -8,29 +10,61 @@ import {
   TextAreaWordLimit,
 } from "./TextAreaModule.styled";
 
-const TextAreaModule = ({
-  title,
-  placeholderContent,
-  maxWords,
-  taregtRef,
-  textAreaContent,
-  handleTextAreaContent,
-  targetInputFocused,
-  setTargetInputFocused,
-}) => {
+const TextAreaModule = ({ title, placeholderContent, maxWords }) => {
+  const textAreaRef = useRef(null);
+
+  const [textAreaFocused, setTextAreaFocused] = useState(false);
+  const [textAreaContent, setTextAreaContent] = useState("");
+
+  useEffect(() => {
+    const handleClickOutside = (e) => {
+      if (
+        textAreaRef.current &&
+        !textAreaRef.current.contains(e.target) &&
+        e.target !== textAreaRef.current
+      ) {
+        setTextAreaFocused(false);
+      }
+    };
+
+    const handleInputFocusBlur = (e) => {
+      if (e.target === textAreaRef.current) {
+        setTextAreaFocused(e.type === "focus");
+      }
+    };
+
+    document.addEventListener("mousedown", handleClickOutside);
+    document.addEventListener("focus", handleInputFocusBlur, true);
+    document.addEventListener("blur", handleInputFocusBlur, true);
+
+    return () => {
+      document.addEventListener("mousedown", handleClickOutside);
+      document.addEventListener("focus", handleInputFocusBlur, true);
+      document.addEventListener("blur", handleInputFocusBlur, true);
+    };
+  }, [textAreaRef]);
+
+  const handleTextAreaContent = (e) => {
+    const { value } = e.target;
+
+    if (value.length <= maxWords) {
+      setTextAreaContent(value);
+    }
+  };
+
   return (
     <TextAreaModuleContainer>
       <TextAreaSubTitle>{title}</TextAreaSubTitle>
       <TextAreaWordLimit>
-        <TextAreaContainer $isFocused={targetInputFocused}>
+        <TextAreaContainer $isFocused={textAreaFocused}>
           <TextArea
             type="text"
             placeholder={placeholderContent}
             maxLength={maxWords}
-            ref={taregtRef}
+            ref={textAreaRef}
             value={textAreaContent}
             onChange={handleTextAreaContent}
-            onFocus={() => setTargetInputFocused(true)}
+            onFocus={() => setTextAreaFocused(true)}
           />
         </TextAreaContainer>
         <TextAreaCharacterCount>
